# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StockShield AI - Stock fraud detection and analysis API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watchlist:
    get:
      operationId: listWatchlist
      tags: [watchlist]
      summary: Get user watchlist
      responses:
        "200":
          description: Watchlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add a stock to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
  /watchlist/{id}:
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove a stock from watchlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
...
  schemas:
...
    WatchlistItem:
      type: object
      properties:
        id:
          type: integer
        ticker:
          type: string
        stockName:
          type: string
        addedAt:
          type: string
          format: date-time
        riskScore:
          type: number
        currentPrice:
          type: number
        change:
          type: number
      required:
        - id
        - ticker
        - stockName
        - addedAt
        - riskScore
        - currentPrice
        - change
    WatchlistInput:
      type: object
      properties:
        ticker:
          type: string
        stockName:
          type: string
      required:
        - ticker
        - stockName
...